modificado o jdk

+++ java-14/src/main/java/br/com/codenation/StatisticUtil.java
@@ -18,7 +18,7 @@ public class StatisticUtil {
 
 		average = total / elements.length;
 
-		return average;
+		return (int)average;
 	}
 
 	public static int mode(int[] elements) {
